refactor: shared card seleect components
